requirements.txt
click
jinja2
pyyaml
snakemake
